+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `bundle download clear` command to remove downloads (alias as `clean` too)
 - Support `os` option for `download` DSL keyword in `Downloadfile`
 - `bundle download --all-operating-systems` to download files for all OS'es
+- `bundle download help` to show instructions (with `usage` as an alias)
 
 ## 1.0.0